[Fix] [Done] [Parth] : Hoisting of the socketId listener from WhiteboadWrapper to LandingPage
